requirements.txt
crcmod
jsonpath
lxml
numpy
numpy-quaternion
pexpect-serial
pyserial
requests-unixsocket
xmodem
